changes to readme: * fix trailing whitespaces and leading tabs * remove all mention of git * removed -a
I am sure using git was well-meant, but not
only is it unnecessary for this use-case, but
also confusing because we are working with an example
taken from a mercurial repository, and it would be far
more natural to use hg versioning. Since your example
creates a manual backup anyway, even novice users should
be safe when replicating with their own code.

Removing -a because it would be too tempting for novice
users to copy this without thinking about it.