complete overhaul of catkinize_cmakelists Basically the line-by-line approach was limiting the power. Instead this does a snippet by snippet approach, where we only look at snippets like:   foo (...) with newlines and indentation allowed in many places. Then we match the name of the function call.
Also further extended the matches. Basically removing "rosbuild_" in all places is a very bad idea,
since for migration this prefix is ideal for users to find where they have to fix something using a migration guide.
So if the script can handle conversion, it should do so, but not blindly replace.