Simplify make_package_lines().
Leave generate_messages() as comment if no messages are in the package.
This makes it easier to add them later if needed.

PS:
In general one should distinguish between getting the data and building the
result string. A 'template' for the entire CMakeList.txt would be cleaner
than the current approach.